% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_sum_formula}
\alias{compute_sum_formula}
\title{Return sum formula for provided element table.}
\usage{
compute_sum_formula(et)
}
\description{
Return sum formula for provided element table.
}
